Lw/zkmoney fixes (#310)
* Return latest non empty tx hash.

* Show progress indicator if sync from rollup 0.

* Update open graph tags.

* Handle rejection when connecting via wallet connect.

* adjust normal social image

* Change default urls for mainnet.

* Don’t delete a user when there's active action.

+++ zk-money/src/app/user_account.ts
@@ -297,7 +297,7 @@ export class UserAccount extends EventEmitter {
       const prevTxs = await this.getJoinSplitTxs(prevUserId);
       if (prevTxs.some(tx => !tx.settled)) {
         userJoinSplitTxs.push(...migratingTxs.concat(prevTxs));
-      } else if (this.accountState.settled) {
+      } else if (this.accountState.settled && !this.activeAction) {
         await this.db.removeMigratingTxs(this.userId);
         await this.sdk.removeUser(prevUserId);
       }
